# catalog-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for a catalog service.",
        "title": "Catalog Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3003",
    "basePath": "/",
    "paths": {
...
        "/api/v1/catalog/bom/{bom_id}": {
            "delete": {
                "description": "Delete a BOM item by its ID",
                "tags": [
                    "bom"
                ],
                "summary": "Delete a BOM item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BOM Item ID",
                        "name": "bom_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/catalog/menus/{id}/bom": {
            "get": {
                "description": "Get the Bill of Materials for a specific menu item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bom"
                ],
                "summary": "Get BOM by Menu Item ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BOMItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Add an ingredient or sub-menu item to a menu item's BOM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bom"
                ],
                "summary": "Add a BOM item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "BOM Item Data",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.addBOMRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.BOMItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/catalog/menus/{id}/bom/flat": {
            "get": {
                "description": "Recursively expands all sub-recipes and returns a flat list of raw ingredients",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bom"
                ],
                "summary": "Get Flat BOM by Menu Item ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/service.FlatBOMItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.addBOMRequest": {
            "type": "object",
            "properties": {
                "ingredient_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "sub_menu_item_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.BOMItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ingredient": {
                    "$ref": "#/definitions/model.Ingredient"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "menu_item_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "sub_menu_item": {
                    "$ref": "#/definitions/model.MenuItem"
                },
                "sub_menu_item_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Ingredient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "description": "e.g. \"g\", \"ml\", \"piece\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.MenuItem": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_available": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "service.FlatBOMItem": {
            "type": "object",
            "properties": {
                "ingredient_id": {
                    "type": "integer"
                },
                "ingredient_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        }
    }
}